fixed all internal warnings

+++ DearPyGui/src/core/AppItems/custom/mvColorMapScale.cpp
@@ -41,7 +41,7 @@ namespace Marvel {
         mvImGuiThemeScope scope(this);
         mvFontScope fscope(this);
 
-        ImPlot::ColormapScale(m_label.c_str(), m_scale_min, m_scale_max, ImVec2(m_width, m_height), *m_value);
+        ImPlot::ColormapScale(m_label.c_str(), m_scale_min, m_scale_max, ImVec2((float)m_width, (float)m_height), *m_value);
     }
 
     void mvColorMapScale::setExtraConfigDict(PyObject* dict)
@@ -49,8 +49,8 @@ namespace Marvel {
         if (dict == nullptr)
             return;
 
-        if (PyObject* item = PyDict_GetItemString(dict, "min_scale")) m_scale_min = ToFloat(item);
-        if (PyObject* item = PyDict_GetItemString(dict, "max_scale")) m_scale_max = ToFloat(item);
+        if (PyObject* item = PyDict_GetItemString(dict, "min_scale")) m_scale_min = (double)ToFloat(item);
+        if (PyObject* item = PyDict_GetItemString(dict, "max_scale")) m_scale_max = (double)ToFloat(item);
     }
 
     void mvColorMapScale::getExtraConfigDict(PyObject* dict)
